fix: executable path in hooks
On platforms, where soft-serve executable gets installed in a path that
is not the same throughout installation, or upgrades, hooks break. Some
examples of such a situation is when moving soft-serve repositories from
one OS to another OS, or on NixOS/Guix

This commit passes the path to the current executable as an environment
variable $SOFT_SERVE_BIN_PATH to the hooks.

To fix/update existing repository hooks:

  [/path/to/soft-serve/repos] $ grep -ERl \
   '^[^[:space:]]+soft-serve[^[:space:]]+' . | xargs sed -r -e \
   's,^([^[:space:]]*soft-serve[^[:space:]]*),"${SOFT_SERVE_BIN_PATH}",g' -i

Signed-off-by: Ashish SHUKLA <ashish.is@lostca.se>
Co-authored-by: Ayman Bagabas <ayman.bagabas@gmail.com>