fix(mirror): skip LFS instead of failing to mirror a repo
If an LFS client can't be created because it's not compatible for
some reason, skip it entirely for a mirrored repo. This prevents one
possible crash when mirroring fails. It still seems to crash when
accessing user info, so something is still missing. Either way, it
seems valid to skip LFS if it just won't work.