fix(hooks): pre and post hooks now receive the full input line
This fixes an issue where the pre and post hooks won't receive
the full input lines when they're split. This is because the
scanner is used to read the input line by line, and the scanner
strips the newline character from the input. This change adds
the newline character back to the input line before passing it
to the pre and post hooks.

Fixes: https://github.com/charmbracelet/soft-serve/issues/680