fix(ui): help menu on file list view (#719)
* fix(ui): Add additional key function indicators

Add the h and l indications in the help message when
those keys select items from lists.

* fix(ui); Fix action long help on file views

Prior to this commit, the copy command help was not overwritten as
intended, and blame and toggle line number options were shown in help on
the file list view page, but only functioned on the file content view
page.